feat: Change the "purpose" of EntityLabelEvent to no longer do label matching
After the Wynncraft 2.1 update, labels should only be matched to TextDisplays, but since some parts of Wynn still use armor stands, it's useful to keep this event around.
