+++ common/src/main/java/com/wynntils/models/token/TokenModel.java
@@ -8,8 +8,7 @@ import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Model;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
-import com.wynntils.handlers.labels.event.EntityLabelChangedEvent;
-import com.wynntils.handlers.labels.event.EntityLabelVisibilityEvent;
+import com.wynntils.handlers.labels.event.EntityLabelEvent;
 import com.wynntils.mc.event.RemoveEntitiesEvent;
 import com.wynntils.models.inventory.InventoryWatcher;
 import com.wynntils.models.items.items.game.MiscItem;
@@ -76,7 +75,7 @@ public class TokenModel extends Model {
     }
 
     @SubscribeEvent
-    public void onLabelChange(EntityLabelChangedEvent event) {
+    public void onLabelChange(EntityLabelEvent.Changed event) {
         if (!(event.getEntity() instanceof ArmorStand)) return;
 
         StyledText name = event.getName();
@@ -167,7 +166,7 @@ public class TokenModel extends Model {
     }
 
     @SubscribeEvent
-    public void onLabelVisibility(EntityLabelVisibilityEvent event) {
+    public void onLabelVisibility(EntityLabelEvent.Visibility event) {
         if (!event.getVisibility()) {
             // This is the normal way in which gatekeepers are "removed" when done
             int id = event.getEntity().getId();
